pyproject.toml
[tool.poetry]
name = "mithratercera"
version = "0.3.0a"
description = "Discord bot for CHUNITHM-NET International"
authors = ["beerpiss <lacvtg.a1.2023@gmail.com>"]
license = "0BSD"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.2"
aiohttp = "^3.8.6"
yarl = "^1.9.2"
beautifulsoup4 = "^4.12.2"
Pillow = "^10.1.0"
pydantic = "^2.4.2"
json5 = "^0.9.14"
python-dotenv = "^1.0.0"
tortoise-orm = "^0.20.0"
fuzzywuzzy = "^0.18.0"
faust-cchardet = { version = "^2.1.19", optional = true }
aiodns = { version = "^3.1.1", optional = true }
brotli = { version = "^1.1.0", optional = true }
lxml = { version = "^4.9.3", optional = true }
orjson = { version = "^3.9.10", optional = true }
winloop = { version = "^0.0.9", optional = true, markers = "sys_platform == 'win32'" }
uvloop = { version = "^0.19.0", optional = true, markers = "sys_platform == 'win32'" }
tzdata = { version = "^2023.3", markers = "sys_platform == 'win32'" }
python-levenshtein = { version = "^0.23.0", optional = true }
seqlog = { version = "^0.3.30", optional = true }

[tool.poetry.extras]
speedups = ["faust-cchardet", "aiodns", "brotli", "lxml", "orjson", "uvloop", "winloop", "python-levenshtein"]

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
types-beautifulsoup4 = "^4.12.0.7"
ruff = "^0.1.3"
pyright = "1.1.333"

[tool.isort]
profile = "black"

[tool.ruff]
select = [
    "A",
    "ANN",
    "ASYNC",
    "ARG",
    "B",
    "BLE",
    "COM",
    "CPY",
    "C4",
    "C90",
    "DTZ",
    "E",
    "EM",
    "F",
    "FBT",
    "FIX",
    "FLY",
    "FURB",
    "ISC",
    "LOG",
    "N",
    "Q",
    "PERF",
    "PGH",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "RET",
    "RSE",
    "RUF",
    "SIM",
    "SLF",
    "SLOT",
    "TCH",
    "TID",
    "TRY",
    "TID252",
    "T10",
    "UP",
    "YTT",
    "W",
]

ignore = [
    # line-too-long (E501)
    # Nothing we can really do if a line is too long, without hurting the readability
    # by breaking up the line into non-sensible parts.
    "E501", 
    # missing-trailing-comma (COM812)
    # Black handles the formatting.
    "COM812"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"